+++ Source/Libraries/GSF.Historian/Files/StateRecordDataPoint.cs
@@ -200,7 +200,7 @@ namespace GSF.Historian.Files
             if (length >= FixedLength)
             {
                 // Binary image has sufficient data.
-                Time = new TimeTag(LittleEndian.ToDouble(buffer, startIndex));
+                Time = new TimeTag((decimal)LittleEndian.ToDouble(buffer, startIndex));
                 Flags = LittleEndian.ToInt32(buffer, startIndex + 8);
                 Value = LittleEndian.ToSingle(buffer, startIndex + 12);
